requirements.txt
absl-py==1.0.0
Flask==2.0.3
Flask-Cors==3.0.9
Flask-RESTful>0.3.8
google-cloud-documentai==1.2.0
google-auth==2.6.0
pylint==2.12.2
gunicorn==20.1.0
